refactor: Move spectrum to elements folder
