+++ plugins/ui/src/js/src/elements/spectrum/RangeSlider.tsx
