+++ plugins/ui/src/js/src/elements/spectrum/Slider.tsx
